NFC: Move third-party headers into _vendor
cocot includes header files from different third-party sources:

* The VPI headers from the SystemVerilog LRM.
* The VHPI headers from the VHPI/VHDL specification.

Move these files into the `_vendor` directory to indicate that those
files are not files cocotb "owns", and as such might have different
formatting, different licenses, etc.

On the change to vpi_user.h: sv_vpi_user.h includes vpi_user.h, both
are part of the SV LRM and located in the same directory. Instead of
searching for vpi_user.h in the system include path, specify the include
path relative to the project root to ensure we get the right file as
vendored into cocotb.
