+++ .pre-commit-config.yaml
@@ -2,12 +2,12 @@
 # Licensed under the Revised BSD License, see LICENSE for details.
 # SPDX-License-Identifier: BSD-3-Clause
 
+exclude: "^cocotb/_vendor"
 repos:
 - repo: "https://github.com/psf/black"
   rev: "22.3.0"
   hooks:
   - id: "black"
-    exclude: "^cocotb/_vendor"
     args:
     - "."
 
@@ -15,7 +15,6 @@ repos:
   rev: "5.10.1"
   hooks:
   - id: "isort"
-    exclude: "^cocotb/_vendor"
     args:
     - "--profile=black"
     - "."
@@ -24,7 +23,6 @@ repos:
   rev: "3.9.2"
   hooks:
   - id: "flake8"
-    exclude: "^cocotb/_vendor"
 
 - repo: "https://github.com/pre-commit/mirrors-clang-format"
   rev: "v13.0.0"
