+++ cocotb/_vendor/README.md
@@ -1,0 +1,14 @@
+# Third-party code
+
+This directory contains code from various third parties, which is distributed
+together with cocotb. Note that some code may be licensed differently from
+cocotb itself; refer to the individual files for details.
+
+## SystemVerilog VPI
+
+The headers `vpi/vpi_user.h` and `vpi/sv_vpi_user.sv` are part of the
+SystemVerilog LRM.
+
+## VHPI
+
+The header `vhpi/vhpi_user.h` is part of the VHPI/VHDL standard.
